requirements.txt
asttokens==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pywin32==310
pyzmq==26.3.0
setuptools==78.1.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
torch==2.6.0
torch-vision==0.1.6.dev0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.0
wcwidth==0.2.13
